requirements.txt
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.0
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
bandit==1.7.5
cffi==1.15.1
coverage==7.2.5
decorator==5.1.1
Django==4.2.1
django-cors-headers==4.0.0
django-coverage-plugin==3.0.0
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-extensions==3.2.1
django-hosts==5.2
django-redis==5.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
execnet==1.9.0
executing==1.2.0
factory-boy==3.2.1
Faker==18.9.0
flake8==6.0.0
gitdb==4.0.10
GitPython==3.1.31
iniconfig==2.0.0
ipython==8.7.0
isort==5.12.0
jedi==0.18.2
markdown-it-py==2.2.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
packaging==23.1
parso==0.8.3
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.7.0
pytest==7.3.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-env==0.8.1
pytest-xdist==3.3.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.5.5
rich==13.3.5
six==1.16.0
smmap==5.0.0
sqlparse==0.4.4
stack-data==0.6.2
stevedore==5.1.0
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.6.0
wcwidth==0.2.6
